CC3XX: Consistent validation of the private key in ECDSA
Make sure that the private key is consistently validated
both in the getpub() and sign() APIs of the lowlevel driver.
Add a common function which loads in the PKA and validates
against limits based on the loaded curve parameters. Also,
make sure that the length of the private_key is validated
initially only against the register_size, as that step makes
only sure that the private key can be loaded on the PKA.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Ia75e3008352f1b1382b150f2f68ab73f049e97a3