CC3XX: validate_point() must set the modulus first
validate_point() performs validation of an elliptic curve
point to belong to the desired curve, when DFA countermeasures
are enabled. Before performing modular arithmetics it must make
sure that it sets correctly the modulus to the field size to
avoid ending up using the curve order as modulus, then restore
the state of the modulus register as it found it.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I0c9f423df6841cf3d8f9108a5cddd9224bd003e6