CC3XX: Tests: Only create test target when enabled
Reuse the existing TEST_CC3XX CMake variable, but instead of using it in
the source code, change it to only compile the CC3XX test code when the
variable is defined. We then have ifdefs in the places that call into
the test code, preventing any usages of the functions when the CMake
variable is not defined.

In the context of the RSE tests, preventing the CC3XX tests target being
linked, prevents the main CC3XX target also being linked. This is an
issue because the main CC3XX target also links in the platform_bl1_1
CMake target which contains all of the BL1_1 source code. For the RSE,
we want to link in symbols from the ROM instead of the RSE tests binary
creating their own copy.

Change-Id: I33eaf78bd08a1e827f617c53e55b21a31d85719f
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>